refactor: each command has its own dir
